requirements.txt
# ── Core ───────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic-settings>=2.2.0

# ── Diffusion / ML ─────────────────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.28.0
controlnet-aux>=0.0.7
xformers>=0.0.24           # comment out on CPU-only systems

# ── Computer Vision ─────────────────────────────────────────────────────────────
opencv-python-headless>=4.9.0.80
Pillow>=10.2.0
mediapipe>=0.10.0
scipy>=1.12.0
numpy>=1.26.0

# ── Queue ──────────────────────────────────────────────────────────────────────
celery[redis]>=5.3.0
redis>=5.0.0

# ── Storage ────────────────────────────────────────────────────────────────────
boto3>=1.34.0
